Update clear.js
it works now but only for Admins

+++ commands/clear.js
@@ -1,10 +1,7 @@
-/*const message = require("../events/guild/message")
-
 module.exports = {
     name: 'clear',
     aliases: ['prune', 'purge'],
-    //permissions: ["ADMINISTRATOR"],
-    botOwner: ['120392991662800896'],
+    permissions: ["ADMINISTRATOR"],
     description: 'Clear messages.',
     async execute(message, args, cmd, client, Discord){
         if (!args[0]) return message.reply('Enter the amount of messages you want to delete');
@@ -13,9 +10,11 @@ module.exports = {
         if (args[0] > 100) return message.reply('Cant do more than 100 messages, sorry!');
         if (args[0] < 1) return message.reply('You must delete at least one message!');
 
-        await message.channel.messages.fetch({limit: args[0]}).then(messages =>{
-            message.channel.bulkDelete(messages);
-        })
-
+        if (!message.member.hasPermission("ADMINISTRATOR")) return message.reply('You do not permission to do this');
+        if (message.member.hasPermission("ADMINISTRATOR")){
+            await message.channel.messages.fetch({limit: args[0]}).then(messages =>{
+                message.channel.bulkDelete(messages);
+            })
+        }
     }
-}*/
+}
